requirements.txt
flask==3.0.3
tensorflow==2.15.0         # or tensorflow-cpu==2.15.0 on CPU-only hosts
mediapipe==0.10.14
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
gunicorn==21.2.0
